+++ src/components/home/recommendation/CategoryList.vue
@@ -1,0 +1,7 @@
+<template></template>
+
+<script>
+export default {};
+</script>
+
+<style></style>
